Tweak to r6747 to allow the user to enter non-integer scale values, and to have scales larger than 1:2^31
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@6754 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/app/qgisapp.h
@@ -175,7 +175,7 @@ public slots:
   //copy the click coord to clipboard and let the user know its there
   void showCapturePointCoordinate(QgsPoint &);
   //! Slot to show current map scale;
-  void showScale(long theScale);
+  void showScale(double theScale);
   //! Slot to handle user scale input;
   void userScale();
   //! Remove a layer from the map and legend
